reorganize cmd folder and rename command class
